I need help with Import
When I opened my code today I was disgusted by my object with the winning conditions. So I decided to try and import it from another document so I dont have to see it between my code. But since I put the import command in my code the function to spin does not work anymore. Error 'spin' is not declared. Pls help